Bugfix:
  1. Fix the problem that the new password under the newly deployed
  cluster etc is inconsistent with the old password
  2. Fix ODP unable to handle configurations larger than 4K, currently
  there is no limit to configuration size
  3. Fix the cluster without LDG. After the enable_ldg switch is turned
  on, the proxy stops updating the location cache, resulting in no sense
  of offline machines.
  4. Fix Under the 2.0 protocol, the data returned to the driver by the ps
  and fetch protocols is incorrect, resulting in the client hang problem
  5. Fix mutli swichover obproxy core at
  databuff_print_key_obj<oceanbase::common::ObLogPrintPointerCntx
  6. Fix obproxy memory leak caused by frequent push of current
  limiting rules
  7. Fix inaccurate routing in col = ("123") mode
  8. Fix too many current-limiting rules causing the push to fail to
  take effect
  9. Fix obproxy memory leak caused by frequent push of current
  limiting rules
  10. Fix After repairing the failover of the standby database into the main
  database, after the original main database is down, the proxy will
  have a short time to route the request to the original main database
  11. Fix Failed to establish connection without specifying cluster id
  within 3 minutes after lossless FO
  12. Fix After repairing the deletion of the standby database, continue to
  send requests to the old clusterId
  13. Fix the problem that the new protocol link proxy cursor scene,
  when the fench reaches the second value, it will be rammed
  14. Fix the problem of ps cursor ramming under non-compression
  protocol
  15. Fix a new connection reporting too many connections error
  16. Fix the case-insensitive name field of the ssl_config
  configuration table, resulting in unreasonable configuration reading
  logic
  17. Fix Compression and 2.0 protocol, obdump import data will be hung
  18. Fix the new protocol, refcursor cannot get the value when there
  are other out parameters
  19. Fix obproxy core at oceanbase::common::ObObjCmpFuncs::compare null
  safe
  20. Fix obproxy core at
  oceanbase::common::ObTimeConverter::str_to_digit_with_date
  21. Fix the error -5560, Unknown prepared statement handle when bmsql
  opens ps when the compression protocol is enabled

Link: https://code.aone.alibaba-inc.com/oceanbase-ce-publish/obproxy/codereview/8205314

+++ src/lib/alloc/object_set.cpp
@@ -384,7 +384,7 @@ void ObjectSet::free_object(AObject *obj)
   if (NULL == bm_
       && NULL == free_lists_
       && normal_hold_bytes_ > ABLOCK_SIZE
-      && (normal_used_bytes_
+      && (static_cast<double>(normal_used_bytes_)
           < static_cast<double>(normal_hold_bytes_)
           * (1. - FREE_LISTS_BUILD_RATIO))) {
     build_free_lists();
